Move ResultOrderer to own component
