+++ src/components/ResultOrderer/index.js
@@ -1,7 +1,7 @@
 import { withStyles } from '@material-ui/core';
 import { connect } from 'react-redux';
 import { injectIntl } from 'react-intl';
-import { setDirection, setOrder } from '../../../redux/actions/sort';
+import { setDirection, setOrder } from '../../redux/actions/sort';
 import ResultOrderer from './ResultOrderer';
 import styles from './styles';
