+++ src/components/TabLists/TabLists.js
@@ -8,7 +8,7 @@ import { intlShape } from 'react-intl';
 import isClient, { parseSearchParams, stringifySearchParams, AddEventListener } from '../../utils';
 import ResultList from '../Lists/ResultList';
 import PaginationComponent from '../PaginationComponent';
-import ResultOrderer from './ResultOrderer';
+import ResultOrderer from '../ResultOrderer';
 import config from '../../../config';
 
 class TabLists extends React.Component {
